refactor: 팀 삭제 시 리더인지 체크 후 예외를 서비스 계층에서 발생

+++ src/main/java/homeTry/team/service/TeamService.java
@@ -111,7 +111,8 @@ public class TeamService {
         Team team = teamRepository.findById(teamId)
                 .orElseThrow(TeamNotFoundException::new);
 
-        team.validateIsLeader(member.getId()); //팀 리더인지 체크
+        if (!team.validateIsLeader(member.getId())) //팀 리더인지 체크
+            throw new NotTeamLeaderException();
 
         teamMemberMappingService.deleteAllTeamMemberFromTeam(team); // 해당 팀에 대한 TeamMemberMapping 데이터 삭제
